Make _link_to_case2_output not a staticmethod
This is needed to record the call in the log

This also is cleaner, and is possible now that we have infrastructure for
setting up a fake case and fake test

Also, to support this refactor, I have moved CaseFake to its own module, and the
tests of this class to their own module.

Test suite: python -m unittest discover
Test baseline: N/A
Test namelist changes: N/A
Test status: N/A

Fixes: None

User interface changes?: No

Code review: None
