+++ utils/python/CIME/SystemTests/tests/test_system_tests_compare_two_clone.py
@@ -9,133 +9,10 @@ from collections import namedtuple
 import os
 import shutil
 import tempfile
-from copy import deepcopy
 
 from CIME.SystemTests.system_tests_compare_two_clone import SystemTestsCompareTwoClone
 import CIME.test_status as test_status
-
-# ========================================================================
-# Fake version of Case object that provides the functionality needed for these
-# tests
-#
-# TODO(wjs, 2016-08-10) Should this be moved into case.py? (Is it useful enough
-# to warrant that?)
-# ========================================================================
-
-class CaseFake(object):
-    def __init__(self, case_root, create_case_root=True):
-        """
-        Initialize a new case object for the given case_root directory.
-
-        Args:
-            case_root (str): path to CASEROOT
-            create_case_root (bool): If True, creates the directory given by case_root
-        """
-        self.vars = dict()
-        if create_case_root:
-            os.makedirs(case_root)
-        self.set_value('CASEROOT', case_root)
-        casename = os.path.basename(case_root)
-        self.set_value('CASE', casename)
-        self.set_value('CASEBASEID', casename)
-        self._set_rundir()
-
-    def get_value(self, item):
-        """
-        Get the value of the given item
-
-        Returns None if item isn't set for this case
-
-        Args:
-            item (str): variable of interest
-        """
-        return self.vars.get(item)
-
-    def set_value(self, item, value):
-        """
-        Set the value of the given item to the given value
-
-        Args:
-            item (str): variable of interest
-            value (any type): new value for item
-        """
-        self.vars[item] = value
-
-    def copy(self, newcasename, newcaseroot):
-        """
-        Create and return a copy of self, but with CASE and CASEBASEID set to newcasename,
-        CASEROOT set to newcaseroot, and RUNDIR set appropriately.
-
-        Args:
-            newcasename (str): new value for CASE
-            newcaseroot (str): new value for CASEROOT
-        """
-        newcase = deepcopy(self)
-        newcase.set_value('CASE', newcasename)
-        newcase.set_value('CASEBASEID', newcasename)
-        newcase.set_value('CASEROOT', newcaseroot)
-        newcase._set_rundir()
-
-        return newcase
-
-    def create_clone(self, newcase, keepexe=False):
-        """
-        Create a clone of the current case. Also creates the CASEROOT directory
-        for the clone case (given by newcase).
-
-        Args:
-            newcase (str): full path to the new case. This directory should not
-                already exist; it will be created
-            keepexe (bool, optional): Ignored
-
-        Returns the clone case object
-        """
-        newcaseroot = os.path.abspath(newcase)
-        newcasename = os.path.basename(newcase)
-        os.makedirs(newcaseroot)
-        clone = self.copy(newcasename = newcasename, newcaseroot = newcaseroot)
-
-        return clone
-
-    def flush(self):
-        pass
-
-    def _set_rundir(self):
-        """
-        Assumes CASEROOT is already set; sets an appropriate RUNDIR (nested
-        inside CASEROOT)
-        """
-        self.set_value('RUNDIR', os.path.join(self.get_value('CASEROOT'), 'run'))
-
-# ========================================================================
-# Tests of CaseFake
-# ========================================================================
-
-class TestCaseFake(unittest.TestCase):
-
-    def setUp(self):
-        self.tempdir = tempfile.mkdtemp()
-
-    def tearDown(self):
-        shutil.rmtree(self.tempdir, ignore_errors=True)
-
-    def test_create_clone(self):
-        # Setup
-        old_caseroot = os.path.join(self.tempdir, 'oldcase')
-        oldcase = CaseFake(old_caseroot)
-        oldcase.set_value('foo', 'bar')
-
-        # Exercise
-        new_caseroot = os.path.join(self.tempdir, 'newcase')
-        clone = oldcase.create_clone(new_caseroot)
-
-        # Verify
-        self.assertEqual('bar', clone.get_value('foo'))
-        self.assertEqual('newcase', clone.get_value('CASE'))
-        self.assertEqual('newcase', clone.get_value('CASEBASEID'))
-        self.assertEqual(new_caseroot, clone.get_value('CASEROOT'))
-        self.assertEqual(os.path.join(new_caseroot, 'run'),
-                         clone.get_value('RUNDIR'))
+from CIME.SystemTests.tests.case_fake import CaseFake
 
 # ========================================================================
 # Structure for storing information about calls made to methods
@@ -207,6 +84,8 @@ class SystemTestsCompareTwoFake(SystemTestsCompareTwoClone):
 
         Args:
             case1 (CaseFake): existing case
+            run_one_suffix (str, optional): Suffix used for first run. Defaults
+                to 'base'. Currently MUST be 'base'.
             run_two_suffix (str, optional): Suffix used for the second run. Defaults to 'test'.
             case2setup_raises_exception (bool, optional): If True, then the call
                 to _case_two_setup will raise an exception. Defaults to False.
@@ -297,10 +176,7 @@ class SystemTestsCompareTwoFake(SystemTestsCompareTwoClone):
         """
         return CaseFake(caseroot, create_case_root=False)
 
-    @staticmethod
-    def _link_to_case2_output(casename1, casename2,
-                              rundir1, rundir2,
-                              run2suffix):
+    def _link_to_case2_output(self):
         # FIXME(wjs, 2016-08-10) remove pass, uncomment this
         pass
         """
